Add files via upload
Project- Our motivation to build this project:
• Offer advice to novice Kaggle users on how to select a competition depending on their level
of competence.
• Provide competition creators with a description of what makes their content successful on
this platform and offer them a sense of assessing how well their new competition will do among
Kaggle users.
With this project, we hope to use information about competitions that was scraped from
https://www.kaggle.com to find answers to two key questions.
Research question 1: Do the Kaggle competitions have varying levels of difficulty? How many
and what elements go into these hardness levels? How many different levels of difficulty may the
contests be divided into. Examine whether this classification is assisting brand-new Kaggle users
in acclimating to the system.
Research question 2: Is the popularity of the organization that posts a competition the only
factor in the competition’s success? Or is it the product of several factors? Determine all such
characteristics that make a competition successful on Kaggle.